// src/nodes/gen/mask.ts
// GENERATED by comfy-workflows codegen — do not edit.
// objectInfoHash: a765285a89bbcdfc58054e715699443e9c764fdc41ca83512c89e2e4177baed3
// Regenerate with: cwf codegen

import { bool, combo, conn, defineNode, float, int, str } from "../../builder/types.js";

export const GrowMask = defineNode("GrowMask", {
  category: "mask",
  inputs: {
    "mask": conn("MASK"),
    "expand": int({ default: 0, min: -18446744073709552000, max: 18446744073709552000, step: 1 }),
    "tapered_corners": bool({ default: true }),
  },
  outputs: [{ name: "MASK", type: "MASK" }] as const,
});

export const LoadImageMask = defineNode("LoadImageMask", {
  category: "mask",
  inputs: {
    "mask": combo(["example.png", "mask.png"]),
    "channel": combo(["alpha", "red", "green", "blue"]),
  },
  outputs: [{ name: "MASK", type: "MASK" }] as const,
});