// src/nodes/gen/image.ts
// GENERATED by comfy-workflows codegen — do not edit.
// objectInfoHash: a765285a89bbcdfc58054e715699443e9c764fdc41ca83512c89e2e4177baed3
// Regenerate with: cwf codegen

import { bool, combo, conn, defineNode, float, int, str } from "../../builder/types.js";

export const ImageScale = defineNode("ImageScale", {
  category: "image/upscaling",
  inputs: {
    "upscale_method": combo(["nearest-exact", "bilinear", "area", "bicubic", "lanczos"]),
    "width": int({ default: 512, min: 0, max: 18446744073709552000, step: 1 }),
    "height": int({ default: 512, min: 0, max: 18446744073709552000, step: 1 }),
    "crop": combo(["disabled", "center"]),
    "image": conn("IMAGE"),
  },
  outputs: [{ name: "IMAGE", type: "IMAGE" }] as const,
});

export const ImageUpscaleWithModel = defineNode("ImageUpscaleWithModel", {
  category: "image/upscaling",
  inputs: {
    "upscale_model": conn("UPSCALE_MODEL"),
    "image": conn("IMAGE"),
  },
  outputs: [{ name: "IMAGE", type: "IMAGE" }] as const,
});

export const LoadImage = defineNode("LoadImage", {
  category: "image",
  inputs: {
    "image": combo(["example.png", "photo.jpg", "mask.png"]),
  },
  outputs: [{ name: "IMAGE", type: "IMAGE" }, { name: "MASK", type: "MASK" }] as const,
});

export const PadImageForOutpainting = defineNode("PadImageForOutpainting", {
  category: "image/inpaint",
  inputs: {
    "image": conn("IMAGE"),
    "left": int({ default: 0, min: 0, max: 18446744073709552000, step: 8 }),
    "top": int({ default: 0, min: 0, max: 18446744073709552000, step: 8 }),
    "right": int({ default: 0, min: 0, max: 18446744073709552000, step: 8 }),
    "bottom": int({ default: 0, min: 0, max: 18446744073709552000, step: 8 }),
    "feathering": int({ default: 20, min: 0, max: 18446744073709552000, step: 8 }),
  },
  outputs: [{ name: "IMAGE", type: "IMAGE" }, { name: "MASK", type: "MASK" }] as const,
});

export const PreviewImage = defineNode("PreviewImage", {
  outputNode: true,
  category: "image",
  inputs: {
    "images": conn("IMAGE"),
  },
  outputs: [] as const,
});

export const SaveImage = defineNode("SaveImage", {
  outputNode: true,
  category: "image",
  inputs: {
    "images": conn("IMAGE"),
    "filename_prefix": str({ default: "ComfyUI" }),
  },
  outputs: [] as const,
});